Added some code changes for some small enhancements to the application
Added the ability to create the factories by listing them in the
properties file.

Added a log4j.xml file to configure a console appender

Added a shutdown hook to perform any cleanup of the engine.

Added changes to the build to package the files into a zip file and
generate a shell script to execute the application.