Global config from the hub provides defaults that can be overridden
Made code changes to take the global configuration from the hub and use
that as the defaults for the properties for the email extension.  If the
same properties exist in the properties file local to the extension the
settings local to the extension take precedence and overwrite the values
that were provided with the global configuration from the Hub.