Refactor the location of some classes to avoid confusion with Restlet
Refactor where the EmailEngine is located.  Rename the Router objects in
the notifier.routers package to include Notifier in the class name.
This avoids confusion with the Restlet Routers in the extension
packages.  Rename any variables using router in the classes and packages
that aren't in the extension package to avoid confusion.