Fix the performance unit test to generate notifications
Because the created date was not unique the wrong number of notification
events were getting generated.  And because the hashcode ended up
becoming the same the policy override and cleared events were not being
added properly to the notification TreeSet for testing.  Therefore need
to ensure the created time for the notification is unique in order to
create the correct expected sets for the tests.